Add search functionality
Added ability to query for users by uid, email, first/last name,
type and company.

For string data, it is possible to search for matches that START
with the query. Substring searches are not currently supported for
performance reasons (full table scan!). Also overhauled column
verification for the users table.